requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
chromadb
google-generativeai
python-dotenv
langchain-text-splitters
